Extract shared SendAction component from send-row (#191)
Refactor-only prep for the Orders-page action stack (#192). Pulls the send
mechanics + sent-state display out of send-row into a reusable SendAction:
sms: deep-link, MWS extension bridge postMessage, clipboard fallback,
recordOptimistic + recordSend wiring, Send/Re-send button, status/copied/error.

- send-action.tsx (new): SendAction with a `revalidate` prop (Send page →
  /admin/send; Orders → /admin/orders) and `onSentChange` callback.
- send-row.tsx: slimmed to the row layout (name/phone) + <SendAction>; mirrors
  sent-state onto <li data-sent> so the test/CSS contract is unchanged.
- record-send.ts: optional 4th `revalidate` param (default "/admin/send") so
  the 3-arg call sites stay unchanged.

No behavior change — existing send + notifications Playwright specs pass as the
parity gate (desktop 8/8; mobile send path green).
